Prepare to remove ProgramOne and ProgramTwo
