+++ src/main/java/main/resources/GetTimelineResource.java
@@ -6,6 +6,9 @@ import main.ProgramTwo;
 import com.codahale.metrics.annotation.Timed;
 
 import twitter4j.Status;
+import twitter4j.Twitter;
+import twitter4j.TwitterException;
+import twitter4j.TwitterFactory;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -26,11 +29,15 @@ public class GetTimelineResource {
         this.counter = new AtomicLong();
     }
 
+	/*
+	 * How to use: //TODO: make this more readable
+	 * curl -i -X GET http://localhost:8080/api/1.0/twitter/timeline
+	 */
     @GET
     @Timed
     public Timeline getTweets() { //TODO: add error checking
     	List<String> value = new LinkedList<>();
-    	List<Status> statuses = ProgramTwo.getHomeTimelineStatuses();
+    	List<Status> statuses = getHomeTimelineStatuses();
     	if (statuses == null){
     		//TODO: exception
     	}
@@ -41,8 +48,18 @@ public class GetTimelineResource {
     }
 
 	/*
-	 * How to use: //TODO: make this more readable
-	 * curl -i -X GET http://localhost:8080/api/1.0/twitter/timeline
+	 * Returns List of Status objects from Home Timeline
+	 * Returns null if TwitterException occurs when trying to retrieve statuses
 	 */
+	public static List<Status> getHomeTimelineStatuses() {
+		Twitter twitter = TwitterFactory.getSingleton();
+		try {
+			List<Status> statuses = twitter.getHomeTimeline();
+			return statuses;
+		} catch (TwitterException e) {
+			e.printStackTrace();
+			return null;
+		}
+	}
 
 }
